+++ lib/stock_markets/data_processing/network_calls.rb
@@ -1,0 +1,3 @@
+module StockMarkets
+  class 
+end
